backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
openai==1.109.1
python-dotenv==1.1.1
pydantic==2.12.3
